requirements.txt
Logbook==0.6.0
configs==2.0.7
docopt==0.6.1
fluent-logger==0.3.3
